Update COMMUNITY_GUIDELINES.md
@uiboy Thank you!

@bardurc Good idea, I've added the asterisk added before "On harmful beliefs".

@lifenautjoe Sorry about my HTML tags. Full disclosure, this is my first time with Markdown (and GitHub) :), but not with HTML. I was used to making anchor links that way (either with the name or id attributes), and I had checked to make sure it was applicable with Markdown, as shown here:

https://stackoverflow.com/questions/6695439/how-to-link-to-a-named-anchor-in-multimarkdown

So if HTML tags can't be used in the app... what should we do about the anchor links? I just tested them, and they still don't work for me; when clicked on from within the document, they do nothing, and when they are accessed from outside the app, they just point at the top of the guidelines page:(.

I'm committing the asterisk change to the patch-2 branch, I hope this comment goes after the ones already posted by @shantanu, @bardurc and @lifenautjoe.